requirements.txt
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
streamlit>=1.28.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# File processing
pypdf>=4.0.0
python-docx>=1.1.0

# Optional: for enhanced NLP
# torch>=2.0.0
# transformers>=4.36.0

